[test_loop] Simple TestLoopEnvBuilder
